Move some items in `util.context` to appropriate places
context_manager was only used once in airflow-core, so I redefined
it where it was used. Likewise context_copy_partial. Then Context was
added to the depreccation for easy import
