+++ providers/cncf/kubernetes/src/airflow/providers/cncf/kubernetes/version_compat.py
@@ -38,9 +38,11 @@ AIRFLOW_V_3_1_PLUS = get_base_airflow_version_tuple() >= (3, 1, 0)
 if AIRFLOW_V_3_1_PLUS:
     from airflow.models.xcom import XCOM_RETURN_KEY
     from airflow.sdk import BaseHook, BaseOperator
+    from airflow.sdk.definitions.context import context_merge
 else:
     from airflow.hooks.base import BaseHook  # type: ignore[attr-defined,no-redef]
     from airflow.models import BaseOperator
+    from airflow.utils.context import context_merge  # type: ignore[attr-defined, no-redef]
     from airflow.utils.xcom import XCOM_RETURN_KEY  # type: ignore[no-redef]
 
 if AIRFLOW_V_3_0_PLUS:
@@ -64,4 +66,5 @@ __all__ = [
     "TaskDecorator",
     "task_decorator_factory",
     "XCOM_RETURN_KEY",
+    "context_merge",
 ]
