+++ providers/standard/src/airflow/providers/standard/decorators/bash.py
@@ -33,8 +33,8 @@ else:
     )
 
 from airflow.providers.standard.operators.bash import BashOperator
+from airflow.providers.standard.version_compat import context_merge
 from airflow.sdk.definitions._internal.types import SET_DURING_EXECUTION
-from airflow.utils.context import context_merge
 from airflow.utils.operator_helpers import determine_kwargs
 
 if TYPE_CHECKING:
