+++ providers/standard/src/airflow/providers/standard/operators/python.py
@@ -51,9 +51,8 @@ from airflow.exceptions import (
 from airflow.models.variable import Variable
 from airflow.providers.standard.hooks.package_index import PackageIndexHook
 from airflow.providers.standard.utils.python_virtualenv import prepare_virtualenv, write_python_script
-from airflow.providers.standard.version_compat import AIRFLOW_V_3_0_PLUS, BaseOperator
+from airflow.providers.standard.version_compat import AIRFLOW_V_3_0_PLUS, BaseOperator, context_merge
 from airflow.utils import hashlib_wrapper
-from airflow.utils.context import context_copy_partial, context_merge
 from airflow.utils.file import get_unique_dag_module_name
 from airflow.utils.operator_helpers import KeywordParameters
 from airflow.utils.process_utils import execute_in_subprocess
@@ -79,7 +78,7 @@ if TYPE_CHECKING:
     try:
         from airflow.sdk.definitions.context import Context
     except ImportError:  # TODO: Remove once provider drops support for Airflow 2
-        from airflow.utils.context import Context
+        from airflow.utils.context import Context  # type: ignore[attr-defined, no-redef]
 
     _SerializerTypeDef = Literal["pickle", "cloudpickle", "dill"]
 
@@ -485,9 +484,19 @@ class _BasePythonVirtualenvOperator(PythonOperator, metaclass=ABCMeta):
     def _iter_serializable_context_keys(self):
         pass
 
+    @staticmethod
+    def context_copy_partial(source: Context, keys: Container[str]) -> Context:
+        """
+        Create a context by copying items under selected keys in ``source``.
+
+        :meta private:
+        """
+        new = {k: v for k, v in source.items() if k in keys}
+        return cast("Context", new)
+
     def execute(self, context: Context) -> Any:
         serializable_keys = set(self._iter_serializable_context_keys())
-        serializable_context = context_copy_partial(context, serializable_keys)
+        serializable_context = self.context_copy_partial(context, serializable_keys)
         return super().execute(context=serializable_context)
 
     def get_python_source(self):
