+++ providers/standard/src/airflow/providers/standard/version_compat.py
@@ -40,9 +40,11 @@ AIRFLOW_V_3_1_PLUS: bool = get_base_airflow_version_tuple() >= (3, 1, 0)
 # even though it wasn't used.
 if AIRFLOW_V_3_1_PLUS:
     from airflow.sdk import BaseHook, BaseOperator
+    from airflow.sdk.definitions.context import context_merge
 else:
     from airflow.hooks.base import BaseHook  # type: ignore[attr-defined,no-redef]
     from airflow.models.baseoperator import BaseOperator  # type: ignore[no-redef]
+    from airflow.utils.context import context_merge  # type: ignore[no-redef, attr-defined]
 
 if AIRFLOW_V_3_0_PLUS:
     from airflow.sdk import BaseOperatorLink
@@ -59,4 +61,5 @@ __all__ = [
     "BaseHook",
     "BaseSensorOperator",
     "PokeReturnValue",
+    "context_merge",
 ]
